Speed up layer building in mcts
Instead of rebuilding all the historical layers each time, we now store
and reuse them instead, flipping as necessary. All layers are stored
relative to the current player.

This also adds a nice hook for quickly benchmarking the mcts engine. You
can run a quick benchmark like so:

    python -m cProfile -o prof src/main.py benchmark --model path/to/model

And view the profile data with runsnake (snakeviz on windows?).

There still appear to be issues with mcts/the value head. In one case,
the player had a chance to take a queen for free, and initially rated
the move highly, but after simulating ~10-20k states, the value dropped
substantially. This probably means that there's either a bug in the
alpha/beta pruning, the evaluation, or that mcts is hooking into some
noise in the model somewhere.

Next things to tackle:
* Adding tests for mcts
* Visualizing the decision tree for mcts somehow to assist debugging
* Weaving in the policy head to avoid noise in the value head
* Parallelizing mcts to saturate the GPU better