Reuse mcts tree for subsequent moves turns and split tj engines
We could be smarter about attempting to reuse the tree by storing all
historical mainline nodes instead of just checking if the board matches
the current root. Then if you undid a move or entered a move manually,
we could still reuse the tree by walking through it move by move from
the root, pruning unneeded nodes.

This doesn't speed up searching yet as we still simulate a fixed number
of batches. If we limit the number of states we explore or limit the
amount of time we explore for, we should get faster / better results.